requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asyncer==0.0.10
attrs==25.4.0
autopep8==2.3.2
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.1
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
coloredlogs==15.0.1
cryptography==46.0.3
cssselect2==0.8.0
defusedxml==0.7.1
dnspython==2.8.0
dropbox==12.0.2
et_xmlfile==2.0.0
fastapi==0.119.1
ffmpy==0.6.4
filelock==3.20.0
filetype==1.2.0
Flask==3.0.3
flask-cors==6.0.1
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
gunicorn==22.0.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
ply==3.11
pooch==1.8.2
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyMatting==1.1.14
pymongo==4.15.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
rembg==2.0.67
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.1
safehttpx==0.1.6
scikit-image==0.25.2
scipy==1.16.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
stone==3.3.1
sympy==1.14.0
tifffile==2025.10.16
tinycss2==1.4.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.0.6
yarl==1.22.0
